Introduces PrioritisedOrderedThreadPoolExecutor
This is the initial version of PrioritisedOrderedThreadPoolExecutor, which intends to augment the functionality
provided in OrderedThreadPoolExecutor by allowing to assign priority to certain sessions over others.

The implementation provided here is based on a copy of the implementation of OrderedThreadPoolExecutor, and
introduces a considerable amount of code duplication. This commit primarily serves as a base for discussion of
the feature, and is intended to be followed by a commit that refactors the code base, to reduce code duplication.

We've introduced this functionality after observing issues in a deployment where system resources where being
starved by the sheer amount of sessions that attempted to perform TLS. Without the class introduced by this
commit, events for any session could eventually fail (time-out), causing the session to fail. If that session
happened to be a session that had already established TLS, the resources that were spent on establishing TLS
are wasted. The negative effect is amplified by the fact that a typical client in such a situation would attempt
to reconnect, which further adds to the load of the system already being starved.

With the class introduced here, priority can be given to sessions that have already established TLS. This
dramatically reduces the issue described above, as the first sessions to fail will be those that are still
negotiating TLS. Using a specialized Comparator, one can even prioritize between these, causing sessions for
which least effort has performed to fail before sessions that are more likely to near TLS completion.