Use OrderedThreadPoolExecutor behavior when not comparing.
If no Comparator is provided, the implementation defaults to FIFO behavior for session handling, as
does OrderedThreadPoolExecutor. With this commit, it now also uses the same BlockingQueue implementation,
further reducing functional changes.

This commit intends to modify the implementation of PriorityOrderedThreadPoolExecutor in such a way that
its existence can be refactored in a feature addition to the implementation of OrderedThreadPoolExecutor.